Add table & column stats on desktop
