+++ libs/database-types/src/schema.ts
@@ -22,6 +22,11 @@ export const ColumnValue = z.union([z.string(), z.number(), z.boolean(), z.null(
 export type ColumnRef = { table: TableId, column: ColumnName }
 export const ColumnRef = z.object({table: TableId, column: ColumnName}).strict()
 
+export function parseTableId(id: TableId): { schema: SchemaName, table: TableName } {
+    const parts = id.split('.')
+    return parts.length === 2 ? {schema: parts[0], table: parts[1]} : {schema: '', table: id}
+}
+
 export type AzimuttSchema = { tables: AzimuttTable[], relations: AzimuttRelation[], types?: AzimuttType[] }
 export type AzimuttTable = {
     schema: AzimuttSchemaName,
@@ -54,30 +59,37 @@ export type AzimuttColumnName = string
 export type AzimuttColumnType = string
 export type AzimuttColumnValue = string
 
+export type TableSampleValues = { [column: string]: ColumnValue }
+export const TableSampleValues = z.record(ColumnValue)
+
 // keep sync with backend/lib/azimutt/analyzer/table_stats.ex & frontend/src/Models/Project/TableStats.elm
 export interface TableStats {
     schema: SchemaName | null
     table: TableName
     rows: number
-    sample_values: { [column: string]: ColumnValue }
+    sample_values: TableSampleValues
 }
 
 export const TableStats = z.object({
     schema: SchemaName.nullable(),
     table: TableName,
     rows: z.number(),
-    sample_values: z.record(ColumnValue)
+    sample_values: TableSampleValues
 }).strict()
+
+export type ColumnCommonValue = { value: ColumnValue, count: number }
+export const ColumnCommonValue = z.object({value: ColumnValue, count: z.number()})
 
 // keep sync with backend/lib/azimutt/analyzer/column_stats.ex & frontend/src/Models/Project/ColumnStats.elm
 export interface ColumnStats {
     schema: SchemaName | null
     table: TableName
     column: ColumnName
+    type: ColumnType
     rows: number
     nulls: number
     cardinality: number
-    common_values: { value: ColumnValue, count: number }[]
+    common_values: ColumnCommonValue[]
 }
 
 export const ColumnStats = z.object({
@@ -88,5 +100,5 @@ export const ColumnStats = z.object({
     rows: z.number(),
     nulls: z.number(),
     cardinality: z.number(),
-    common_values: z.object({value: ColumnValue, count: z.number()}).array()
+    common_values: ColumnCommonValue.array()
 }).strict()
