+++ libs/shared/src/index.ts
@@ -18,5 +18,5 @@ export type DesktopBridge = {
     databaseQuery: (url: DatabaseUrl, query: string) => Promise<DatabaseResults>
     databaseSchema: (url: DatabaseUrl) => Promise<AzimuttSchema>
     tableStats: (url: DatabaseUrl, table: TableId) => Promise<TableStats>
-    columnStats: (url: DatabaseUrl, column: ColumnRef) => Promise<ColumnStats>
+    columnStats: (url: DatabaseUrl, ref: ColumnRef) => Promise<ColumnStats>
 }
