+++ libs/utils/src/object.ts
@@ -1,8 +1,8 @@
-export function mapValues<T, U>(obj: Record<string, T>, f: (v: T) => U): Record<string, U> {
+export function mapValues<T, U>(obj: Record<string, T>, f: (t: T) => U): Record<string, U> {
     return Object.fromEntries(Object.entries(obj).map(([k, v]) => [k, f(v)]))
 }
 
-export function filterValues<T>(obj: Record<string, T>, p: (v: T) => boolean): Record<string, T> {
+export function filterValues<T>(obj: Record<string, T>, p: (t: T) => boolean): Record<string, T> {
     return Object.fromEntries(Object.entries(obj).filter(([, v]) => p(v)))
 }
